Refactoriza historial con callbacks de métricas
