+++ tests/test_history.py
@@ -2,14 +2,14 @@
 
 import pytest
 
-from tnfr.dynamics import _update_history
+from tnfr.metrics import _metrics_step
 
 
 def test_phase_sync_and_kuramoto_recorded(graph_canon):
     G = graph_canon()
     G.add_node(1, theta=0.0)
     G.add_node(2, theta=0.0)
-    _update_history(G)
+    _metrics_step(G)
     hist = G.graph.get("history", {})
     assert hist["phase_sync"][-1] == pytest.approx(1.0)
     assert "kuramoto_R" in hist
