+++ tests/test_validators.py
@@ -58,6 +58,7 @@ def test_validator_glyph_invalido():
     G = _base_graph()
     n0 = list(G.nodes())[0]
     set_attr_str(G.nodes[n0], ALIAS_EPI_KIND, "INVALID")
+    G.nodes[n0]["glyph_history"] = ["INVALID"]
     with pytest.raises(ValueError):
         run_validators(G)
